Probing the framework: PROB/
The framework can be used to probe (latin: probare) the framework.

Under PROB/ is placed a pseudo project that is exercizing the features
of the framework.

PROB/conf/RKBANC.map has an extensive explanation about the map syntax
and the way it drives compilation, verifications and installation.

PROB/ has to be kept up-to-date with the features offered by the
framework.