librkcompsh: add rk_find_cmd_regex()
Add the ability to search a command, in the PATH given, using a legacy
regex (ex. '^python[0-9.]*$'). This will print all the matching
commands. Have to filter, for example for versioning, after.

This is obviously introduced for python even if support for
pre-compilation of modules has to be done at installation done,
without requiring to know the exact version present on a node,
allowing a same package to be installed on the TARGET specified, but
without defining, in advance, what exact version of python is here
(only the minimum version required to interpret the scripts).