[F] Adding auxiliary install cmds: python
Auxiliary commands can be added by defining RKINSTALL_AUX_CMDS with
the name of an existing POSIX.2 shell script (name.cmds). These files
are placed in the install_bin/ subdir of the install hier, allowing
to post-process, on the TARGET, files.

Here, for example, a python.cmds is added so that byte-compiled
version of *.py source files can be generated, on the TARGET, at
install time.

Rationale: cross-compilation has sense since one does not need a
compiler to run the program. But for python, the byte-compilation is
done by the interpreter that is required to run, on the TARGET, the
python code; furthermore, the python byte-compiled code is not
portable. So this has typically to be done on TARGET, without
requiring a running version of whatever python on the MATRIX.