aux/rkpython: a script to search for a matching version python.
This is an utility to find in the PATH a python interpreter optionally
matching a version condition written as "op" "major" ["minor" ["micro"]]
<op> being "<" | "<=" | "==" | ">=" | ">". The best i.e. greatest
matching version instance found is returned on stdout, with an exit
status of 0 in case of success; nothing and an exit status of 1 if
syntax error, of exit status of 2 if nothing matching found. Can be
used without requiring a version.