aux/rkpython: return both path and subdir to install modules.
Third parties modules are installed in various locations. The norm
is to put them in 'lib/pythonMajor.Minor/site-packages' (the prefix
is not specified and will vary depending on where the user wants to
install _these_---making python find them is another problem).

So we return both informations, separated by a space, on stdout.