rkbuild: circumvent dash(1) bug with IFS.
Field splitting doesn't seem to work in some circumstances with (some)
version of dash(1), in the present context in a subshell call: a
variable defined with two arguments separated by a tab, with the
following calling expression (variable not in double quotes):

echo $line

is passed with the tab i.e. as if field splitting was not done---one
argument to echo as if line was expanded in double quotes.

Fix: in the regex, test for tabs as well as for spaces.

Problem reported by B. Atticus Grobe.